+++ packages/react-ui-codemodes/font-to-svg/__testfixtures__/button_comp.output.js
@@ -1,0 +1,3 @@
+import Button from "retail-ui/Button";
+
+props => <Button icon={props.icon} />;