+++ packages/react-ui-codemodes/font-to-svg/__testfixtures__/icon_empty.output.js
@@ -1,0 +1,3 @@
+import Icon from "@skbkontur/react-icons";
+
+() => <Icon />;