+++ packages/react-ui-codemodes/font-to-svg/__testfixtures__/icon_exp_ok.input.js
@@ -1,0 +1,3 @@
+import Icon from "retail-ui/Icon";
+
+() => <Icon name={"Ok"} />;