+++ packages/retail-ui/components/ComboBoxOld/ComboBoxOld.js
@@ -1,3 +1,4 @@
+// @flow
 /* eslint-disable flowtype/no-types-missing-file-annotation */
 /* eslint-disable flowtype/no-weak-types */
 import * as React from 'react';