+++ packages/retail-ui/components/Modal/__stories__/Modal.stories.tsx
@@ -5,9 +5,9 @@ import { storiesOf } from '@storybook/react';
 import Modal from '../';
 import Button from '../../Button';
 import Input from '../../Input';
-import Textarea from '../../Textarea';
 import Toggle from '../../Toggle';
 import Upgrades from '../../../lib/Upgrades';
+import BorderAllIcon from '@skbkontur/react-icons/BorderAll';
 
 const basicFontStyle = {
   fontSize: '14px',
@@ -101,7 +101,7 @@ class ModalWithScrollableContent extends Component<
   };
 }
 
-class ModalWithInputInHeader extends Component<{}, { opened: boolean }> {
+class ModalWithIconInput extends Component<{}, { opened: boolean }> {
   public state = {
     opened: false
   };
@@ -110,17 +110,65 @@ class ModalWithInputInHeader extends Component<{}, { opened: boolean }> {
     return (
       <Modal onClose={this.close}>
         <Modal.Header>
-          <Input placeholder="Some input placeholder..." />{' '}
-          <Input size="large" placeholder="Some large input placeholder..." />
-          <br />
-          <Textarea placeholder="Some textarea placeholder" value="" />
+          <Input
+            size="large"
+            placeholder="Modal.Header"
+            leftIcon={<BorderAllIcon />}
+            rightIcon={<BorderAllIcon />}
+          />
+          <Input
+            size="medium"
+            placeholder="Modal.Header"
+            leftIcon={<BorderAllIcon />}
+            rightIcon={<BorderAllIcon />}
+          />
+          <Input
+            size="small"
+            placeholder="Modal.Header"
+            leftIcon={<BorderAllIcon />}
+            rightIcon={<BorderAllIcon />}
+          />
         </Modal.Header>
         <Modal.Body>
-          <p>
-            A lotta people ask me where the fuck I've been at the last few
-            years.
-          </p>
+          <Input
+            size="large"
+            placeholder="Modal.Body"
+            leftIcon={<BorderAllIcon />}
+            rightIcon={<BorderAllIcon />}
+          />
+          <Input
+            size="medium"
+            placeholder="Modal.Body"
+            leftIcon={<BorderAllIcon />}
+            rightIcon={<BorderAllIcon />}
+          />
+          <Input
+            size="small"
+            placeholder="Modal.Body"
+            leftIcon={<BorderAllIcon />}
+            rightIcon={<BorderAllIcon />}
+          />
         </Modal.Body>
+        <Modal.Footer>
+          <Input
+            size="large"
+            placeholder="Modal.Footer"
+            leftIcon={<BorderAllIcon />}
+            rightIcon={<BorderAllIcon />}
+          />
+          <Input
+            size="medium"
+            placeholder="Modal.Footer"
+            leftIcon={<BorderAllIcon />}
+            rightIcon={<BorderAllIcon />}
+          />
+          <Input
+            size="small"
+            placeholder="Modal.Footer"
+            leftIcon={<BorderAllIcon />}
+            rightIcon={<BorderAllIcon />}
+          />
+        </Modal.Footer>
       </Modal>
     );
   }
@@ -432,7 +480,7 @@ class ModalWithVariableHeight extends Component<
 
 storiesOf('Modal', module)
   .add('With scrollable parent content', () => <ModalWithScrollableContent />)
-  .add('With Input in header', () => <ModalWithInputInHeader />)
+  .add('With Icon Input', () => <ModalWithIconInput />)
   .add('Modal over another modal', () => <ModalOverAnotherModal />)
   .add('Disabled modal', () => (
     <Modal disableClose>